snapshots: Allow CreateSnapshot RPC to work on a published Nexus
CreateSnapshot RPC tries to obtain a read/write BdevHandle but that
fails on a published Nexus as the "NVMe-oF Target module" has already
claimed it.

As the create snapshot NVMe vendor command only needs a read-only
descriptor, create a copy of spdk_bdev_nvme_io_passthru in the
spdk-sys layer with the check removed so that a read-only BdevHandle
can be used, allowing CreateSnapshot to work.