config: Import pools using the lvs implementation
This makes it consistent with the destroy path, addressing the panic
in #454. If the disks field does not specify a bdev type, default to
uring if the kernel supports it, falling back to aio otherwise.

Shoehorn a regression test for destroying a pool imported via config
in test_snapshot. Fix up the existing tests to cope with the io_if
field being ignored by the Lvs::create_or_import.