Add message bus library
Common api to be used by both mayastor and rust control plane services.

The Message trait is expected to be used to easily publish/request
messages with default channel and message bus.

The Publish/Request traits are more flexible to allow the less typical
use of the non default channel and specify a particular message bus.
Added traits can be implemented for other types by using exported macros

All the different message request/replies are kept in this library for
simplicity.

No timeout/retry mechanism is provided yet so request/reply may hang
if the receiver does not reply back...