Add control plane services library
This library is specifically for running services over the message bus.
It uses a builder pattern allowing for multiple service subscribers
to be registered.
Each ServiceSubscriber must implement a handlerFn which will receive the
message as an argument as well as a filterFn to filter in messages with
the desired message identifiers.

Added example service to showcase how to use the service library