snapshots: Implement create snapshots for local replicas
Local replicas connected over loopback have an 'lvol' bdev, which does
not support NVMe Admin commands so explicitly allow it for any bdev so
that the call makes it as far as nexus_bdev::nvme_admin. From there,
have a separate codepath for lvol bdevs that creates the snapshot.

This requires a separate Nexus::io_completion_local for this case
where there is no actual child IO so that the parent IO records the
status and completes correctly.

Improve error logging and introduce value-only versions of the
cdw10/11 helpers to tidy up the snapshot time retrieval.