core: rework main.rs
A lot of the "main" code is done with in the start closure of
mayastor. This PR moves this into main.rs where it should be. The
older .start() method is still present as they are used by tests but
over time we can phase it out.

The need for this - is mostly to avoid spawning runtimes during init