uring: Re-enable bdev and pool tests
The tests can be re-enabled as Ubuntu 20.04 has updated its kernel
5.4.0-50 which includes the fix for a NULL pointer dereference in
io_sq_wq_submit_work(). The regression was originally introudced in
5.4.0-45, the corresponding upstream fix and first regressed versions
are 5.4.62 and 5.4.51, respectively.

Since the tests were last run, SPDK was upgraded to 20.07 and the
uring bdev module no longer uses IORING_SETUP_IOPOLL so remove the
restrictions on O_DIRECT and filesystem type from bdev::util::uring
and the uring-support CLI.

Also bump liburing to 0.7 and the io-uring crate to 0.4.0.