test: need ability to (easily) drive IO to a nexus
This change adds the ability to create workloads within a given mayastor
instance. This is for testing -- not just during mayastor development but
also for e2e tests where we can invoke job creations form within k8s in
the near future.

An example is given `io_job.rs` where we create 2 nvmf targets through
the compose framework, and then use `struct MayastorTest` to do
some IO to it.