Ensure bdevs are removed when not in use
Bdevs are now opened using the spdk_bdev_open_ext() function (note: SPDK
has deprecated spdk_bdev_open). This allows a callback to be registered
which is called when an asynchronous bdev event, such as removal,
occurs.

When a NexusChild is closed, destroy is called, resulting in spdk
raising a SPDK_BDEV_EVENT_REMOVE event. This calls back to the
NexusChild remove() function which is responsible for closing everything
down in an orderly fashion before the underlying bdev is removed.

Additional changes:

- The RebuildJob no longer stores the source and destination handles.
These are acquired as and when they are needed. This ensures that a bdev
can be removed even while there is an active rebuild process.

- cancel_child_rebuild_jobs() now waits for all rebuild jobs to
terminate before returning

- The reconfigure test case has been removed as it is no longer valid
given that a bdev is removed when setting a child offline.

- The replica_timeout test is now ignored. It still needs to be
determined whether or not this is a valid test case.