Add initial node service...
which listens on the registration channel for mayastor instances and
keeps a record of them.

It starts a watchdog with a set timeout for
each instance and deems a node as offline if the deadline is not met.

Reports this information via the message bus, which is to be consumed
by the rest service.

Added compose test.