Cancel all rebuild jobs on nexus destruction
When destroying a nexus all rebuild jobs should be cancelled before
closing the children. This ensures that the close on the children can
complete properly as any handles held by the rebuild job will no longer
exist.

Previously, destroying the nexus resulted in the rebuild jobs being
stopped. However, this just issued a stop request but did not wait for
the rebuild job to actually stop.

When removing the nexus child, set the bdev to None before calling
reconfigure. This should prevent multiple calls to remove for the same
child.

Note: This change attempts to resolve the issues seen on CI where rebuild
test cases have been observed to hang.