env: add -l option to specify cores
The -l option serves the same purpose as the -m option however it accepts
a list of cores with ranges i.e 1,2,10-15.

Also two default values have been updated while here as the number of
shared buffers was to low to test this change and some notes to the YAML
file has been added.