Try to fix CI nbd issues
Run the set timeout in another thread while we poll as it seems that
when the nbd device is open it triggers a partition scan.
Set the NBD size to 0 before disconnecting as it was observed that when
we disconnect sometimes the device gets rescanned and the kernel thinks
it should partition scan again... setting the size to 0 before the
disconnect seems to alleviate this.