Make nexus children wait for bdev removal on close
A channel is used to allow a nexus child to wait on the removal of the
underlying bdev before completing the close call.

The previous assumption had been that a call to destroy() would only
return when the bdev had actually been removed. However, in some cases,
such as for NVMf bdevs, this was not the case.

When closing a local (loopback) child, the underlying bdev must not be
destroyed. Doing so would result in the lvol and data being deleted.
Instead, the child is only removed, but the underlying bdev remains
untouched.