snapshot: Reintroduce replica snapshot cargo test
Refactor the test to use docker compose so that the remote mayastor
instance runs in a separate container instead of juggling multiple
mayastor instances, which became impossible when ms_exec was removed
in 36680604.

Have Bio::assess allow non-success replies with NVMe status code of
InvalidOpcode to not fault the child as these are only sent by tests.
While I'm here, move the handling of non-success replies to a separate
non-inlined function for readability and to keep the hot path simple.