Move pipeline abort to global scope again.
Now when default github notifications are disabled in jenkins
configuration we can abort the job at global level. Another reason
is that we don't have to enable execution of unsafe code in the
pipeline then. We have already enabled raw job object access but we
would need more to get rid of the latest error message:

 "Scripts not permitted to use method hudson.model.Run getExecutor"

The init stage remains there because it avoids duplicating the code
for setting pending state on a github commit.