Unschedule Mayastor pod to test replica loss
In order to avoid low-level network manipulation to fault a child,
the new test unschedules the corresponding mayastor replica pod,
so removing it from the IO path, to achieve the same effect.
This allows greater flexibility in where the test can be run
(e.g. CI testing) as the kubernetes API is all that is needed.
Removed use of thread in the tests for better cleanup on failure.