fix(nexus): assertion failure when unpublishing a faulted nexus
The assertion occurs due to the nvmf subsystem transitioning from a
paused to an inactive state, which is forbidden. Always resume the
nexus in child_retire so that it is usually in the active state, which
avoids the assertion when unpublishing it.

When there is only a single faulted child, the nexus itself is in a
faulted state, which means no IO is possible. In nexus_bdev, fail the
IO if all submissions failed, which also includes the case where no
submissions were made. Refactor the post-IO submission to avoid
repetition. Also include the case where no usable child was found when
doing a round-robin for reads. This also fixes CAS-606.

Repurpose the existing cargo test for a replica that is stopped, then
continued, to test the cases above with a single remote replica.