test(rebuild): wait for successful rebuild
CI/CD was showing occasional failures of rebuild tests. This was because
the tests generally only waited a short time (i.e. 2 seconds) for a
rebuild to complete. This wasn't always long enough.

The utility function to wait for successful rebuilds has been modified
so that it continually polls for rebuild progress. As long as some
progress is being made within a given window (5 seconds) it will
continue to wait for the rebuild. After 30 seconds, if the rebuild has
not completed, it will give up waiting for it.