test(rebuild): disable test
Disable the rebuild_dst_disconnect test. This test is occasionally
failing on CI/CD even with a suitably high timeout time. When this
occurs an error message is output indicating that we failed to get IO
channel because we were probably low on memory.

A ticket will be raised to investigate this fully and re-enable the
test.