ci(moac): scale up/down test fails sometimes
The problem was that after we add a new replica to the volume it
does not immediately transition to degraded state (the new replica
usually starts in degraded state because it is not synced) but the
volume state change happens shortly after it. So the test code was
wrong that immediately after adding a child to nexus the state should
be degraded. It has to wait for it.