refactor(moac): drop support for NBD in CSI
NBD does not work well with new nexus feature when nexus is created
only if it is used (mounted). For NBD we need to know which node the
nexus is located and the location must not change throughout the
lifetime of the volume. Apparently that does not work now when nexus
is created on whatever node is available and the location can change
between mounts of the volume.

And as part of that convert csi.js to typescript.

Resolves: CAS-672