feat(control plane): allow images to be built
Provide the ability to build container images for the various control
plane components.

Before performing a nix-build, ensure you are in the ../Mayastor
directory.

e.g.
To build a specific control plane service image such as the node
service:
        nix-build -A services.node-image

To build a development image, add the "-dev" suffix:
        nix-build -A services.node-image-dev

To build all service images:
        nix-build -A services

e.g.
To build a specific operator such as the node operator:
       nix-build -A operators.node-image

To build a development image, add the "-dev" suffix:
       nix-build -A operators.node-image-dev

To build all operator images:
       nix-build -A operators