feat(control plane): allow images to be built
Provide the ability to build container images for the control plane
services and operators.

To build the node service, for example:
	nix-build -A images.node-service
or to build a debug version, add the suffix "-dev":
	nix-build -A images.node-service-dev

To build the node operator, for example:
	nix-build -A images.node-operator
or to build a debug version, add the suffix "-dev":
	nix-build -A images.node-operator-dev

Additional change:

Changed the existing image names to remove the "-image" suffix. So the
command to build the mayastor image now looks like:
	nix-build -A images.mayastor
rather than:
	nix-build -A images.mayastor-image