fix(nexus): replace NexusChild destroying flag with child state
Replace the destroying flag with a new Destroying ChildState and store
the previous state so that it when handling the remove event it can be
restored if the previous state was Open or Faulted due to OutOfSync to
preserve existing behaviour.

Also prevent a child from being opened if it is already in Destroying
state.