fix(nexus): set children to destroying state on shutdown
The existing call to destroy all nexuses on shutdown uses a reference
to the list of nexuses, which is itself mutated when a nexus is
destroyed. This can lead to a segfault on some platforms. Destroying a
nexus also leads to the child states being saved to the config file,
which is unintended.

Set all children to Destroying state instead to address both issues.

Fixes: 3ae2903b ("fix(nexus): allow re-adding child removed by AEN")