revert: remove rust-toolchain
We use nix-shell to define the default version of rust (stable). If
people wish to "bring their own" they can use `--arg norust true`
or use `.envrc`.

Having a default rust-toolchain in the repo prevents this previous
behaviour from working correctly. And would require making changes
to the repo if one wants to override it.

I've updated the docs to reflect this.