chore(rest): remove volume operations
The Volume operations were included as a convenience to the consumer of
the REST API. For example, this allowed a volume to be easily created
without having to worry about creating pools, replicas etc. However,
the consumer of the REST API will typically be an Operator that does
have domain knowledge and should therefore be responsible for
constructing a volume. This makes the existing volume operations
redundant and so they have been removed.