feat(nexus): allow shared replica to be added as local child
Move to a build of SPDK that allows that and change the existing
nexus_multipath cargo test to test that scenario.

This also means a bdev could be shared over multiple protocols so
disallow that for a bdev that is already claimed. Add a message for
the iSCSI Error::CreateTarget for consistency with nvmf.

Rename the cargo test that checks it to nexus_share_test for clarity.