feat(rest): authenticate clients
Provide the ability to authenticate client requests through the use of a
JSON Web Token (JWT). This requires a JSON Web Key (JWK) to be available
for use by the REST service so that it can validate the JWT.

The REST service can be started with the '--no-auth' argument to disable
authentication. This is useful for test cases where authentication is
unnecessary.

The REST service can be started with the '--jwk' argument to enable
authentication. The path to the relevant JWK file must also be provided.

Example authentication files are provided in
../Mayastor/control-plane/rest/authentication for test purposes and
should not be used in production.