fix(nexus): handle partial IO submission
This fixes a potential lost IO during submission leading to a potential use
after free. Also, when we have failed write and sufficient children are
available, we do not fail the while IO.

This PR also moves a lot of functions scattered around several files into one.
All IO related functions are not in nexus_io.rs. This will avoid touching many
files in the near future so it only hurts onces.

Lastly, the error store bits are removed and the admin pass through handler
are not moved over yet as they are very likely to change again with the
new driver implementation and its not be used right now anyway. As a
result of the nvme admin commands not being moved over, snapshots are
also non functional.

The code has been tested with the nvmet binary that creates
a nexus using nvmf existing targets.