refactor: rationalise e2e test structure further
All tests register fail handlers and get reporters.
At the moment each test do these individually.
Replace the individual calls with a single call,
to a helper function which does both.
An added advantage is that this then is the best,
place to execute common actions,
for example to emit log markers.