fix(moac): moac assert when publishing a volume
This happened when putting moac under a bit of stress and issuing
create/publish/unpublish/destroy requests in a quick succession.
It revealed a few race conditions. Basically, publish method was
running in "parallel" (interleaved) with FSA code, so the state of
the volume was changing in two places. That has an easy fix of
putting code that changes the state of the volume from publish
method to FSA itself. fsa() guarding method guarantees that we can
have only one _fsa() running at any given time. So that should fix
the race condition. What publish merely does is set a few volume
state variables that tell FSA what to do and kick off FSA that does
the rest. Notification mechanism for letting the publish method know
when the required changes are done is in place but had to be slightly
revamped to be useable for that.

The same change has been done for unpublish and destroy methods (the
same problem although that it wasn't reproduced so far).

Some of the volume object tests have been moved to volumes test
suite because they require to run the FSA now.

Rename children URIs used in the tests so that they reflect the
node where child is located (as it is in real world).

Use events.EventEmitter with type information where possible.

Resolves: CAS-728