feat(nvmeadm): add support for a few controller options
The behaviour of the controller can be tuned better instead of just
accepting the default values from kernel.

Avoid repeating code for connecting to a target and make the code
more rusty.

This work was originally done to address CAS-499 (Mayastor targets
should set sane io timeouts). As it evolved we decided it's better
to document the behavior and suggest how to avoid it instead of
overriding the defaults. Merging it and improving nvmeadm lib
capabilities does not make any harm.