ci: allow e2e tests to run on arbitrary mayastor source tree MQ-430
I preparation for moving the e2e tests to separate repository,
allow e2e tests to run on arbitrary mayastor repo trees.
The e2e tests will use yaml files in the mayastor source
tree to deploy mayastor as defined for the revision of mayastor
that exists in the mayastor source tree

1. Stop using relative paths to locate files in the directory trees,
instead rely on 2 absolute paths
    1. mayastor source tree root
    2. e2e test tree root
2. Reduce reliance on environment variables (which are transient and not
recorded automatically) in favour of using the configuration entity.
3. Document more clearly where the configuration file is searched for and loaded.