fix(moac): cannot unpublish nexus on a node that is down
This introduces a new volume state "offline" that is used for
volumes that are inaccessible (the node with the nexus is down).
It is important distinction from "faulted" state that has been
used for this case previously.  "faulted" speaks for the data
(data loss), while "offline" speaks for availability of the data
(access). FSA has been fixed to handle this new state properly so
it is possible to:

1. create pod with mounted volume (on a different node than nexus),
2. bring the node with nexus down,
3. delete pod with the mounted volume,
4. recreate the pod with the same volume again.

This gives a chance to manually recover the volume by the user
action in case that the node with nexus never comes back online.

What happens in the background is that a series of CSI calls
(unpublish/publish) brings the volume to the state when it can be
used again. Making this process more smooth without requiring the
manual intervention is out of scope of this ticket.

There is a corner case when the old nexus could come back and write
something to replicas while there is already the a nexus. This case
will be fixed on behalf of CAS-833.

Resolves: CAS-816