fix(moac): failed to provision volume on RKE
Two issues, described in great depth in #772 github ticket, were
encountered by a user while trying to provision mayastor volume on
RKE cluster:

1. Mayastor config file was mounted as a directory instead as a file
2. Unhandled exception in Moac (result of partially created volume)

Both issues are fixed in this commit.

Add 1. There does not seem to be a good reason for mounting config
file instead of just the whole config directory so I have removed
that mount. It's not the first time we see this issue and although it
is probably caused by a bug in k8s/OS it can be easily avoided.