feat(core): add spawn_local() for threads
This removes the local macro we have been using for a while and
replaces it with a submission function. Although this looks rather
pointless -- this enables us to run an async runtime in separate
threads and communication is done through rte_ring.

Additionally this will also allow us to run any functions along side
mayastor that does, for example, health checking etc.

IOW, lets say you have core mask 0x1 -- we will allow tokio to create
its own threads and unaffinitze them, and then use this interface to
spawn futures there who are !Send nor !Sync. The spawn_local() function
added here, returns a futures::oneshot::Receiver<T> which can be awaited
on threads that run somewhere but not on the specified mask.

Clippy started to point out that certain asserts should be written
differently so thats part of this PR too.