feat(csi): support WaitForFirstConsumer for mayastor volumes
We want to support late binding of mayastor volumes because we can
optimize distribution of the volume over the nodes. In particular we
can enforce that the nexus and one of the replicas is created on
the same node as the app when creating the volume.

There is a new storage class parameter "local" that enforces the
locality if set to true (the default in our storage class yaml).
If it is not possible to schedule the app on the same node as nexus
and replica and "local" is enabled, then the pod remains in
unscheduleable state until the user rectifies the situation.

There is one downside of local volumes though. Once the node
affinity is set during the volume creation it cannot be changed
later. Assuming that all nodes holding original replicas of the
volume are removed from the cluster, then the app becomes
unschedulable although that the replicas might have been migrated
to new nodes already.

Convert index.js to typescript to catch more errors in the file and
provide simple moac script that calls index.ts:main().

Resolves: CAS-834