chore: update bindgen and other crates
Update bindgen (0.57 to 0.58.1) and fix the use the of deprecated
functions. It doesn't address the layout issue with
spdk_nvme_feat_async_event_configuration but should address layout
issues with unions on AArch64 which we can build for.

As Cargo.toml is being changed anyway, also update
enum-primitive-derive (0.1 to 0.2), num-traits (0.1 to 0.2) and nix
(0.14, 0.16.1 to 0.20) to reduce the number of different versions of
the same dependent crates and remove dependencies on unused crates.