fix(nexus): remove alias from bdevs that are local
If the alias we try to set, is present, we return true rather then
failing the operation.

If the bdev is destroyed, we remove the URI from the alias.

While debugging this i've also decided to remove the tracing
of yield points as they make the calls to noise and have no
real value