feat(nexus): register an NVMe reservation on adding a child
Implement NexusChild::resv_register() which sends an NVMe
Reservation Register command as an NVMe IO Passthru command via
BlockDeviceHandle::{nvme_resv_register,io_passthru} when adding a
new child or when creating a nexus with children, though only to NVMe
child devices.

Add a Rust integration test, tacked on to nexus_multipath as that
already sets up a remote child, to verify that the key has been
registered.

Use a hard-coded key for now as this just lays the groundwork for
subsequent commits to acquire the reservation.