feat: make replica URI's unique
Adds a UUID as a query parameter to the replica's URI making it uniquely identifiable.
The nexus create will then be able to associate a child with a state, retrieved from etcd,
using the URI, and thus avoid having to open the child and retrieve a GUID from the metadata,
if said child is not in a healthy state.