feat(core): provide handle to external runtime
Mayastor has its own futures which are driven by polling of the "green" threads.
If however, we want to do something which does not need or can run on the
mayastor cores, we can dispatch it to a foreign runtime.

Additionally this also fixes a bug where we where interacting with the nvmf
target from an invalid thread.